pyproject.toml
[project]
name = "tetra-kg-agent"
version = "0.1.0"
description = "Scientific Knowledge Graph Agent for drug discovery hypothesis generation"
readme = "DESIGN.md"
requires-python = ">=3.10"
dependencies = [
    # ML & Graph
    "networkx>=3.4",
    "node2vec>=0.4.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.24.0",
    "gensim>=4.3.0",

    # LLM & Agent
    "google-adk>=0.1.0",
    "google-generativeai>=0.8.0",

    # HTTP & Async
    "httpx>=0.27.0",
    "aiohttp>=3.11.0",

    # Data
    "pandas>=2.2.0",
    "pydantic>=2.10.0",

    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.39.0",
    "pyvis>=0.3.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",

    # Observability & Tracing
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "openinference-instrumentation-google-adk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ml", "clients", "extraction", "agent", "models", "frontend", "observability", "pipeline"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
